netplas/api/v1/schemas.py
import coreapi
import coreschema
from rest_framework.schemas import ManualSchema


HelloSchema = ManualSchema(fields=[
    coreapi.Field(
        "asset",
        location="query",
        required=True,
        schema=coreschema.String()
    ),
])